css/css-fonts/parsing/font-kerning-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Level 3: parsing font-kerning with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-fonts/#font-kerning-prop">
<meta name="assert" content="font-kerning supports the full grammar 'auto | normal | none'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("font-kerning", "auto");
test_valid_value("font-kerning", "normal");
test_valid_value("font-kerning", "none");
</script>
</body>
</html>
